Add the TypeAlignmentService and it's documentation
