+++ src/main/scala/com/sneaksanddata/arcane/cdm_change_feed/services/streaming/processors/CdmGroupingProcessor.scala
@@ -14,7 +14,8 @@ import scala.concurrent.duration.Duration
  * The batch processor implementation that converts a lazy DataBatch to a Chunk of DataRow.
  * @param groupingSettings The grouping settings.
  */
-class CdmGroupingProcessor(groupingSettings: GroupingSettings) extends BatchProcessor[LazyList[DataRow], Chunk[DataRow]]:
+class CdmGroupingProcessor(groupingSettings: GroupingSettings, typeAlignmentProcessor: TypeAlignmentService)
+  extends BatchProcessor[LazyList[DataRow], Chunk[DataRow]]:
 
   /**
    * Processes the incoming data.
@@ -24,49 +25,29 @@ class CdmGroupingProcessor(groupingSettings: GroupingSettings) extends BatchProc
   def process: ZPipeline[Any, Throwable, LazyList[DataRow], Chunk[DataRow]] = ZPipeline
     .map[LazyList[DataRow], Chunk[DataRow]](list => Chunk.fromIterable(list))
     .flattenChunks
-    .map(row => toTypedRow(row))
+    .map(typeAlignmentProcessor.alignTypes)
     .groupedWithin(groupingSettings.rowsPerGroup, groupingSettings.groupingInterval)
-
-  private def toTypedRow(row: DataRow): DataRow = row map { cell =>
-    DataCell(cell.name, cell.Type, convertType(cell.Type, cell.value))
-  }
-
-  private def convertType(arcaneType: ArcaneType, value: Any): Any =
-    value match
-      case None => null
-      case Some(v) => convertSome(arcaneType, v)
-
-  private def convertSome(arcaneType: ArcaneType, value: Any): Any = arcaneType match
-    case LongType => value.toString.toLong
-    case ByteArrayType => value.toString.getBytes
-    case BooleanType => value.toString.toBoolean
-    case StringType => value.toString
-    case DateType  => java.sql.Date.valueOf(value.toString)
-    case TimestampType => null //java.sql.Timestamp.valueOf(value.toString) // TODO
-    case DateTimeOffsetType => java.time.OffsetDateTime.parse(value.toString)
-    case BigDecimalType => BigDecimal(value.toString)
-    case DoubleType => value.toString.toDouble
-    case IntType => value.toString.toInt
-    case FloatType => value.toString.toFloat
-    case ShortType => value.toString.toShort
-    case TimeType => java.sql.Time.valueOf(value.toString)
 
 /**
  * The companion object for the LazyOutputDataProcessor class.
  */
 object CdmGroupingProcessor:
+  
+  type Environment = GroupingSettings 
+    & TypeAlignmentService
 
   /**
    * The ZLayer that creates the LazyOutputDataProcessor.
    */
-  val layer: ZLayer[GroupingSettings, Nothing, CdmGroupingProcessor] =
+  val layer: ZLayer[Environment, Nothing, CdmGroupingProcessor] =
     ZLayer {
       for
         settings <- ZIO.service[GroupingSettings]
-      yield CdmGroupingProcessor(settings)
+        tas <- ZIO.service[TypeAlignmentService]
+      yield CdmGroupingProcessor(settings, tas)
     }
 
-  def apply(groupingSettings: GroupingSettings): CdmGroupingProcessor =
+  def apply(groupingSettings: GroupingSettings, typeAlignmentService: TypeAlignmentService): CdmGroupingProcessor =
     require(groupingSettings.rowsPerGroup > 0, "Rows per group must be greater than 0")
     require(!groupingSettings.groupingInterval.equals(Duration.Zero), "groupingInterval must be greater than 0")
-    new CdmGroupingProcessor(groupingSettings)
+    new CdmGroupingProcessor(groupingSettings, typeAlignmentService)
